TEMP: Add link.sh script which I know won't work
I misunderstood what ENV variables from linked containers tell us.  I thought
that they would tell us what the port on the host was for a given 'docker'
port.  I was wrong.  I'll commit this because it has some stuff in it that
might be useful later.  I'll revert it straight away.