Update README and ./script/run.sh
I'm not going to go down the route of automagically loading HOST IP / PORT
details into etcd.  I thought that I could do it by linking containers and
reading envrionment variables.  This is not the case, the environment
variables only hold the 'docker' ports and IP addresses which I don't think
are accessible between host machines.  I could go down the route of
https://github.com/polvi/docker-reg and give a container access to the
docker socket on the host via a volume mount but I don't like the idea
of giving a container this much power.  I'll therefore have to do this on the
host itself.